+++ demo/src/windows.rs
@@ -24,6 +24,7 @@ pub fn controls(
     ui: &Ui,
     selected_provider: &mut Provider,
     possible_providers: &mut dyn Iterator<Item = &Provider>,
+    http_stats: Option<&walkers::HttpStats>,
     image: &mut ImagesPluginData,
 ) {
     Window::new("Satellite")
@@ -42,6 +43,12 @@ pub fn controls(
                         }
                     });
             });
+
+            if let Some(http_stats) = http_stats {
+                ui.collapsing(format!("{:?} HTTP statistics", selected_provider), |ui| {
+                    ui.label(format!("Requests in progress: {}", http_stats.in_progress));
+                });
+            }
 
             ui.collapsing("Images plugin", |ui| {
                 ui.add(Slider::new(&mut image.angle, 0.0..=360.0).text("Rotate"));